TdvfPkg: Fix the #GP caused by VS2019 compiler optimization bug.
In VS2019 there is an optimization feature which will insert below line
into funtion SetPageTablePoolReadOnly (TdShim/Sec/X64/VirtualMemory.c).
    __asm__("movdqa xmm0, [rip+0xadbd]"); RIP=0xffc8620f.
But movdqa requires operands 16-byte aligned if operands is a memory
operand. Otherwise #GP is triggered.

MSFT admits this is an optimization bug. See below links.
https://developercommunity.visualstudio.com/t/vs2019-165-optimizes-some
-memory-stores-to-use-mov/933914?viewtype=solutions
This optimization bug has not been fixed in VS2019(16.9.2).

To work round this issue, LevelShift/LevelMask/LevelSize are declased
as the global variables instead of local variables. So that they're not
optimized by VS2019 in the wrong way.