OvmfPkg/OvmfPkg.dec: Add the PCDs and GUIDs for VTPM
Add the PCDs and GUIDs to support VTPM.
The PCDs like the below:
  PcdVtpmSecureSessionInfoTableAddr,
  PcdVtpmSecureSessionInfoTableSize.

VTPM needs these to record the session info.
It is the part of the messgae that the spdm sends.

The GUIDs like the below:
 gEdkiiVTpmSecureSpdmSessionInfoHobGuid,
 gPeiVmmSpdmTunnelPpiGuid,
 gEdkiiVmmSpdmTunnelProtocolGuid.

VTPM needs the HobGuid to get sessin info in DXE phase,
and the PPI and protocol to provide the support for TPM2.

Signed-off-by: Min Xu <min.m.xu@intel.com>
Signed-off-by: Sun, Ceping <cepingx.sun@intel.com>