OvmfPkg/VmmSpdmVTpm: Extend VTPM Spdm session info after session setup
The VTPM Spdm session info is the first event in RTMR[3]. Before creating
a VTPM Spdm session, the tdvf checks the RTMR[3] to see whether it's
non-zero, if it is, a VTPM Spdm session has already been started. In that
case, the tdvf won't create a new session.

If the RTMR[3] is zero, the tdvf will extend the VTPM Spdm session info to
the RTMR[3] after the session setup successfully.

Signed-off-by: Min Xu <min.m.xu@intel.com>
Signed-off-by: sunceping <cepingx.sun@intel.com>