ArmVirtPkg: Add Null instance of gEfiRealmApertureManagementProtocolGuid
The ArmCcaIoMmuDxe driver depends on gEfiRealmApertureManagementProtocolGuid,
and produces either gEdkiiIoMmuProtocolGuid or gIoMmuAbsentProtocolGuid.
Other drivers, for example those including QemuFwCfgLibMmio, thus depend
on either the IOMMU or IOMMU Absent protocol.

Since ArmCcaIoMmuDxe only produces gEfiRealmApertureManagementProtocolGuid
when in a Realm at the moment, those drivers cannot be loaded when not
in a Realm, and boot doesn't reach BDS. Add an empty instance of
gEfiRealmApertureManagementProtocolGuid so that ArmCcaIoMmuDxe always
produces something and unblocks other drivers.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Signed-off-by: Dhanus M Lal <Dhanus.MLal@fujitsu.com>